# Conflicts:
#	RoguelightSpeedRun20D/UserSettings/Layouts/default-2022.dwlt

+++ RoguelightSpeedRun20D/Assets/_/minwoo/01.Scripts/01.Item/01.Struct/WeaponData.cs
@@ -6,7 +6,7 @@ public struct WeaponData
     float size;
     float cooltime;
 
-    public WeaponData(int damage, bool isRangeAttack, int size, float cooltime)
+    public WeaponData(int damage, bool isRangeAttack, float size, float cooltime)
     {
         this.damage = damage;
         this.isRangeAttack = isRangeAttack;
